Senior Design/mooove/requirements.txt
awscli==1.10.1
awsebcli==3.7.3
blessed==1.9.5
boto==2.38.0
botocore==1.3.23
cement==2.4.0
colorama==0.3.3
decorator==4.0.9
Django==1.9.1
docker-py==1.1.0
dockerpty==0.3.4
docopt==0.6.2
docutils==0.12
jmespath==0.9.0
mysqlclient==1.3.7
networkx==1.11
numpy==1.11.0
pathspec==0.3.3
pyasn1==0.1.9
python-dateutil==2.4.2
PyYAML==3.11
requests==2.6.2
rsa==3.3
six==1.10.0
texttable==0.8.4
virtualenv==14.0.1
wcwidth==0.1.6
websocket-client==0.35.0
wheel==0.26.0
django-filter==0.13.0
django-storages-redux==1.3.2
djangorestframework==3.3.2
